fix(product_category): Autofill category name when exporting category to backend
Exporting a category to a custom backend by creating a new drupal_binding record from
product_category form view where asking the user to input the category name that will be used on
Drupal. With this change the field will be autofilled with main product category name in order to
avoid user errors.