fix(drupal.product.product): Prevent fail export inventory cron job when the product have not being
There are some situations where we can have a product already set to be exported to a backend but

that for any valid reason never where actually exported so we don't have a valid drupal_id. This

kind of situations prevents the cron job to update inventory properly run thus the inventory is

never updated on drupal.